# Bit-exact version of RGB2Luv_b