… modules/imgproc/test/test_lab.cpp
… modules/imgproc/test/test_precomp.hpp