@vpisarev 2017-07-30
@savuor, thank you! Since we claim that this is bit-exact implementation, we should change our tests as well. They should also use softfloat for initialization and fixed-point in the inner loops. There should be zero difference between reference results and your code in the library.