@savuor 2017-08-21
@vpisarev Done. Now there are tests for bit-exactness that calculate CRC of resulting images.
N.B. This should be merged after PR #9374